requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-groq
langchainhub
python-dotenv
openai
pinecone-client
jupyter
firecrawl-py
pydantic
tiktoken
matplotlib
seaborn
python-dateutil
######Flask#########
gunicorn
Flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
email-validator
aiosmtpd
flask-mail
pyjwt
####### google calendar api########
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
######Celery & Redis#####
celery
redis
flower
######Instagram######
instagrapi
######GitGurdian####
pre-commit
#####MongoDB######
pymongo
#####Encryption####
cryptography
